requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
websockets==12.0
pydantic==2.6.4
pydantic-settings==2.2.1

# Audio processing - the scientific core
praat-parselmouth==0.4.3
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
numpy==1.26.4
scipy==1.13.0

# ML / Training
xgboost==2.0.3
scikit-learn==1.4.1.post1
pandas==2.2.1
joblib==1.3.2
requests==2.31.0
imbalanced-learn==0.12.0

# Database
sqlalchemy==2.0.29
aiosqlite==0.20.0

# Environment
python-dotenv==1.0.1
setuptools==45.0.0

# CORS
httpx==0.27.0
